@@ -21,7 +21,7 @@ long_description = file: README.md
 long_description_content_type = text/markdown
 
 [options]
-python_requires = >= 3.6
+python_requires = >= 3.7
 include_package_data = True
 packages = find:
 install_requires =
@@ -32,6 +32,11 @@ install_requires =
   scipy
   sciencedates
   pymap3d
+package_dir=
+    =src
+
+[options.packages.find]
+where=src
 
 [options.extras_require]
 tests =
